refactor, ifc property export
