+++ AddOns/Speckle/Sources/AddOn/Converter/HostToSpeckle/HostToSpeckleConverter.h
@@ -20,8 +20,9 @@ public:
 	std::string GetApplicationFolder() override;
 	std::string GetHostAppReleaseInfo() override;
 	nlohmann::json GetElementMaterialQuantities(const std::string& elemId) override;
-	nlohmann::json GetElementPropertiesByPropertyType(const std::string& elemId, const ArchicadPropertyType propertyType) override;
-	nlohmann::json GetElementPropertiesByPropertyFilter(const std::string& elemId, const PropertyFilter filter) override;
+	nlohmann::json GetElementPropertiesByPropertyType(const std::string& elemId, const ArchicadPropertyTypeFilter propertyType) override;
+	nlohmann::json GetElementPropertiesByPropertyFilter(const std::string& elemId, const PropertyCollectionFilter filter) override;
+	nlohmann::json GetElementIfcProperties(const std::string& elemId) override;
 	nlohmann::json GetElementProperties(const std::string& elemId) override;
 	WorkingUnits GetWorkingUnits() override;
 	ArchicadObject GetArchicadObject(const std::string& elemId, SendConversionResult& conversionResult) override;
