+++ AddOns/Speckle/Sources/AddOn/Converter/HostToSpeckle/IHostToSpeckleConverter.h
@@ -6,21 +6,7 @@
 #include "WorkingUnits.h"
 #include "ArchicadObject.h"
 #include "SendConversionResult.h"
-
-enum class ArchicadPropertyType
-{
-	UserDefined,
-	FundamentalBuiltIn,
-	UserLevelBuiltIn,
-	All
-};
-
-enum class PropertyFilter
-{
-	GeneralRatings,
-	IFC,
-	Dimensional
-};
+#include "PropertyFilters.h"
 
 class IHostToSpeckleConverter 
 {
@@ -40,8 +26,9 @@ public:
 	virtual std::string GetApplicationFolder() = 0;
 	virtual std::string GetHostAppReleaseInfo() = 0;
 	virtual nlohmann::json GetElementMaterialQuantities(const std::string& elemId) = 0;
-	virtual nlohmann::json GetElementPropertiesByPropertyType(const std::string& elemId, const ArchicadPropertyType propertyType) = 0;
-	virtual nlohmann::json GetElementPropertiesByPropertyFilter(const std::string& elemId, const PropertyFilter filter) = 0;
+	virtual nlohmann::json GetElementPropertiesByPropertyType(const std::string& elemId, const ArchicadPropertyTypeFilter propertyType) = 0;
+	virtual nlohmann::json GetElementPropertiesByPropertyFilter(const std::string& elemId, const PropertyCollectionFilter filter) = 0;
+	virtual nlohmann::json GetElementIfcProperties(const std::string& elemId) = 0;
 	virtual nlohmann::json GetElementProperties(const std::string& elemId) = 0;
 	virtual WorkingUnits GetWorkingUnits() = 0;
 	virtual ArchicadObject GetArchicadObject(const std::string& elemId, SendConversionResult& conversionResult) = 0;
